Renaming and re-factoring models to allow proper connections/references
