+++ models/Type.js
@@ -4,9 +4,9 @@ const { Model, DataTypes } = require('sequelize');
 const sequelize = require('../config/connection.js');
 
 /* Establish Table as extension of Model class */
-class CombatStats extends Model {}
+class Type extends Model {}
 
-CombatStats.init(
+Type.init(
     {
         id: {
             type: DataTypes.INTEGER,
@@ -14,17 +14,16 @@ CombatStats.init(
             primaryKey: true,
             autoIncrement: true,
         },
-        hp: {
-            type: DataTypes.INTEGER,
+        name: {
+            type: DataTypes.STRING,
             allowNull: false,
+            validate: {
+                len: [16],
+            }
         },
-        atk: {
+        weakness: {
+            // Points at self. Junction Table?
             type: DataTypes.INTEGER,
-            allowNull: false,
-        },
-        def: {
-            type: DataTypes.INTEGER,
-            allowNull: false,
         },
     },
     {
@@ -32,15 +31,14 @@ CombatStats.init(
         timestamps: false,
         freezeTableName: true,
         underscored: true,
-        modelName: 'combat_stats',
+        modelName: 'type',
     }
 );
 
-module.exports = CombatStats;
+module.exports = Type;
 
 /**
  * id
- * HP
- * ATK
- * DEF
+ * name
+ * weakness_id (self)
  */
