+++ seeds/seed-creatures.json
@@ -5,8 +5,8 @@
         "brand_id" : 1,
         "type_id" : 1,
         "screenname" : "God",
-        "combat_stats" : 1,
-        "care_stats" : 1,
+        "combatstat_id" : 1,
+        "carestat_id" : 1,
         "exp" : 0,
         "health" : 10
     },
@@ -16,8 +16,8 @@
         "brand_id" : 1,
         "type_id" : 1,
         "screenname" : "Admin",
-        "combat_stats" : 2,
-        "care_stats" : 2,
+        "combatstat_id" : 2,
+        "carestat_id" : 2,
         "exp" : 0,
         "health" : 10
     },
@@ -27,8 +27,8 @@
         "brand_id" : 2,
         "type_id" : 2,
         "screenname" : "davidlucio",
-        "combat_stats" : 3,
-        "care_stats" : 3,
+        "combatstat_id" : 3,
+        "carestat_id" : 3,
         "exp" : 10,
         "health" : 10
     },
@@ -38,8 +38,8 @@
         "brand_id" : 3,
         "type_id" : 3,
         "screenname" : "davidlucio",
-        "combat_stats" : 4,
-        "care_stats" : 4,
+        "combatstat_id" : 4,
+        "carestat_id" : 4,
         "exp" : 0,
         "health" : 10
     }
